chart-extraction/requirements.txt
anthropic
requests
streamlit
pandas
python-dotenv
PyJWT
cryptography
